requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic[email]>=1.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
python-dotenv>=0.19.0
requests 